Cleanup. Deduplicate. Simplify. Slightly document.
